Rename TropoDB index files
