+++ implementation/rocksdb/db/zns_impl/impl/tropodb_impl_background.cc
@@ -17,9 +17,9 @@
 #include "db/write_batch_internal.h"
 #include "db/zns_impl/tropodb_config.h"
 #include "db/zns_impl/tropodb_impl.h"
-#include "db/zns_impl/index/zns_compaction.h"
-#include "db/zns_impl/index/zns_version.h"
-#include "db/zns_impl/index/zns_version_set.h"
+#include "db/zns_impl/index/tropodb_compaction.h"
+#include "db/zns_impl/index/tropodb_version.h"
+#include "db/zns_impl/index/tropodb_version_set.h"
 #include "db/zns_impl/io/szd_port.h"
 #include "db/zns_impl/persistence/zns_manifest.h"
 #include "db/zns_impl/persistence/zns_wal.h"
