+++ implementation/rocksdb/db/zns_impl/index/tropodb_compaction.h
@@ -8,9 +8,9 @@
 
 #include "db/dbformat.h"
 #include "db/zns_impl/tropodb_config.h"
-#include "db/zns_impl/index/zns_version.h"
-#include "db/zns_impl/index/zns_version_edit.h"
-#include "db/zns_impl/index/zns_version_set.h"
+#include "db/zns_impl/index/tropodb_version.h"
+#include "db/zns_impl/index/tropodb_version_edit.h"
+#include "db/zns_impl/index/tropodb_version_set.h"
 #include "db/zns_impl/persistence/zns_manifest.h"
 #include "db/zns_impl/ref_counter.h"
 #include "db/zns_impl/table/zns_sstable_manager.h"
