+++ implementation/rocksdb/db/zns_impl/index/tropodb_version_edit.cc
@@ -1,10 +1,10 @@
 // Copyright (c) 2011 The LevelDB Authors. All rights reserved.
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file. See the AUTHORS file for names of contributors.
-#include "db/zns_impl/index/zns_version_edit.h"
+#include "db/zns_impl/index/tropodb_version_edit.h"
 
 #include "db/zns_impl/tropodb_config.h"
-#include "db/zns_impl/index/zns_version.h"
+#include "db/zns_impl/index/tropodb_version.h"
 #include "db/zns_impl/table/zns_zonemetadata.h"
 #include "db/zns_impl/utils/tropodb_logger.h"
 #include "rocksdb/rocksdb_namespace.h"
